Separate out utxo_store.rs
