+++ crates/subcoin-snapcake/src/state_sync_wrapper.rs
@@ -1,3 +1,4 @@
+use crate::utxo_store::UtxoStore;
 use codec::Decode;
 use sc_client_api::ProofProvider;
 use sc_network_sync::strategy::state_sync::{
@@ -14,107 +15,6 @@ use subcoin_runtime_primitives::Coin;
 use subcoin_utxo_snapshot::{Utxo, UtxoSnapshotGenerator};
 
 const DOWNLOAD_PROGRESS_LOG_INTERVAL: Duration = Duration::from_secs(5);
-
-#[derive(serde::Serialize, serde::Deserialize)]
-struct UtxoCsvEntry {
-    txid: bitcoin::Txid,
-    vout: u32,
-    is_coinbase: bool,
-    amount: u64,
-    height: u32,
-    script_pubkey: String,
-}
-
-impl From<Utxo> for UtxoCsvEntry {
-    fn from(utxo: Utxo) -> Self {
-        let Utxo { txid, vout, coin } = utxo;
-        Self {
-            txid,
-            vout,
-            is_coinbase: coin.is_coinbase,
-            amount: coin.amount,
-            height: coin.height,
-            script_pubkey: hex::encode(&coin.script_pubkey),
-        }
-    }
-}
-
-impl From<UtxoCsvEntry> for Utxo {
-    fn from(csv_entry: UtxoCsvEntry) -> Self {
-        let UtxoCsvEntry {
-            txid,
-            vout,
-            is_coinbase,
-            amount,
-            height,
-            script_pubkey,
-        } = csv_entry;
-
-        let coin = Coin {
-            is_coinbase,
-            amount,
-            height,
-            script_pubkey: hex::decode(script_pubkey).expect("Failed to decode script_pubkey"),
-        };
-
-        Self { txid, vout, coin }
-    }
-}
-
-enum UtxoStore {
-    // Initially created for local testing.
-    #[allow(unused)]
-    InMem(Vec<Utxo>),
-    Csv(PathBuf),
-}
-
-impl UtxoStore {
-    fn push(&mut self, utxo: Utxo) {
-        match self {
-            Self::InMem(list) => list.push(utxo),
-            Self::Csv(path) => {
-                // Open the file in append mode each time to avoid keeping it open across calls
-                let file = std::fs::OpenOptions::new()
-                    .append(true)
-                    .open(path)
-                    .expect("Failed to open file");
-
-                let mut wtr = csv::WriterBuilder::new()
-                    .has_headers(false) // Disable automatic header writing
-                    .from_writer(file);
-
-                let utxo_csv_entry = UtxoCsvEntry::from(utxo);
-                if let Err(e) = wtr.serialize(&utxo_csv_entry) {
-                    panic!("Failed to write UTXO entry to CSV: {}", e);
-                }
-
-                if let Err(e) = wtr.flush() {
-                    panic!("Failed to flush CSV writer: {}", e);
-                }
-            }
-        }
-    }
-
-    fn count(&self) -> std::io::Result<usize> {
-        match self {
-            Self::InMem(list) => Ok(list.len()),
-            Self::Csv(path) => {
-                // Open the file in read mode only for counting
-                let file = std::fs::File::open(path)?;
-                let reader = std::io::BufReader::new(file);
-
-                // Count lines by reading through each record
-                let count = csv::ReaderBuilder::new()
-                    .has_headers(false)
-                    .from_reader(reader)
-                    .records()
-                    .count();
-
-                Ok(count)
-            }
-        }
-    }
-}
 
 /// Wrapped [`StateSync`] to intercept the state response for parsing the UTXO state.
 pub(crate) struct StateSyncWrapper<B: BlockT, Client> {
@@ -155,7 +55,8 @@ where
         let snapshot_filepath = snapshot_dir.join("snapshot.dat");
         let snapshot_file =
             std::fs::File::create(&snapshot_filepath).expect("Failed to create output file");
-        let snapshot_generator = UtxoSnapshotGenerator::new(snapshot_filepath, snapshot_file);
+        let snapshot_generator =
+            UtxoSnapshotGenerator::new(snapshot_filepath, snapshot_file, bitcoin::Network::Bitcoin);
 
         let utxo_filepath = snapshot_dir.join("utxo.csv");
 
@@ -207,7 +108,6 @@ where
 
                     self.muhash.insert(&data);
 
-                    // TODO: write UTXO to a local file instead of storing in memory.
                     self.utxo_store.push(Utxo { txid, vout, coin });
                     self.received_coins += 1;
                 }
@@ -246,40 +146,13 @@ where
                 self.snapshot_generator.path().display()
             );
 
-            match &mut self.utxo_store {
-                UtxoStore::InMem(list) => {
-                    let utxos = std::mem::take(list);
-
-                    self.snapshot_generator
-                        .write_utxo_snapshot(
-                            self.target_bitcoin_block_hash,
-                            utxos_count as u64,
-                            utxos,
-                        )
-                        .expect("Failed to write UTXO set snapshot");
-                }
-                UtxoStore::Csv(path) => {
-                    let file = std::fs::File::open(path).expect("Failed to open utxo.csv");
-
-                    let reader = std::io::BufReader::new(file);
-                    let csv_reader = csv::ReaderBuilder::new()
-                        .has_headers(false)
-                        .from_reader(reader);
-
-                    let utxo_iter = csv_reader
-                        .into_deserialize::<UtxoCsvEntry>()
-                        .filter_map(Result::ok)
-                        .map(Utxo::from);
-
-                    self.snapshot_generator
-                        .write_utxo_snapshot(
-                            self.target_bitcoin_block_hash,
-                            utxos_count as u64,
-                            utxo_iter,
-                        )
-                        .expect("Failed to write UTXO set snapshot");
-                }
-            }
+            self.utxo_store
+                .write_snapshot(
+                    &mut self.snapshot_generator,
+                    self.target_bitcoin_block_hash,
+                    utxos_count,
+                )
+                .expect("Failed to write UTXO set snapshot");
 
             tracing::info!(
                 target: "snapcake",
